+++ Tortuga.Chain/Shared/Tests/Appenders/SqlInfoMessageEventTests.cs
@@ -1,0 +1,38 @@
+﻿namespace Tests.Appenders;
+
+#if SQL_SERVER_MDS
+
+using Microsoft.Data.SqlClient;
+
+using Tests.Models;
+using Tortuga.Chain;
+using Tortuga.Chain.SqlServer;
+
+#endif
+
+#if (SQL_SERVER_MDS)
+
+[TestClass]
+public class SqlInfoMessageEventTests : TestBase
+{
+	[DataTestMethod, RootData(DataSourceGroup.Primary)]
+	public void SqlInfoMessageEventTests_InfoMessageEvents(string dataSourceName)
+	{
+		var dataSource = DataSource(dataSourceName);
+		try
+		{
+			string message = null;
+			SqlInfoMessageEventHandler handler = (object sender, SqlInfoMessageEventArgs e) => { message = e.Message; };
+
+			var result = dataSource.Procedure("HR.EmployeeList").ToTable().WithInfoMessageNotification(handler).Execute();
+
+			Assert.AreEqual("Listing employees", message);
+		}
+		finally
+		{
+			Release(dataSource);
+		}
+	}
+}
+
+#endif